pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codex-compliance-agent"
version = "1.0.0"
description = "Codex - Operational Risk & Compliance Agent"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    # Core
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "click>=8.1.0",
    
    # LLM & Embeddings
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-nvidia-ai-endpoints>=0.1.0",
    "openai>=1.6.0",
    
    # Vector Store & RAG
    "chromadb>=0.4.0",
    "llama-index>=0.10.0",
    "llama-index-embeddings-nvidia>=0.1.0",
    "llama-index-llms-nvidia>=0.1.0",
    "llama-index-vector-stores-chroma>=0.1.0",
    "langchain-chroma>=0.1.0",
    "langchain-text-splitters>=0.0.1",
    
    # Document Processing
    "pypdf>=3.17.0",
    "python-magic>=0.4.27",
    
    # Web UI
    "streamlit>=1.28.0",
    "streamlit-shadcn-ui>=0.1.0",
    
    # API & HTTP
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    
    # Data Processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    
    # Sandbox (safe execution)
    "restrictedpython>=7.0",
    
    # Utilities
    "python-dateutil>=2.8.0",
    "tqdm>=4.66.0",
    "fpdf2>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
    "flake8>=6.1.0",
]

[project.scripts]
codex = "src.ui.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
